requirements.txt
matplotlib==3.7.1
numpy==1.24.3
Pillow==9.5.0
Pillow==9.5.0
streamlit==1.18.1
streamlit_drawable_canvas==0.9.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
